feat: Update format of recordings to be individual files
The single file for recordings.json was too large for the app to process
and too large to commit to the repo.  It's clear we're better off with
individual files.

This commit updates the generate scripts to output individual files and
rebuilds all the generated and search data.  It also builds a new
data.zip which has these files in it.

We have not updated the relevant documentation in the repo yet